// src/gen/api/types/product_catalog_domain_dt_os_category_tree_history_detail_dto.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * ProductCatalog.API
 * OpenAPI spec version: 1.0
 */
import type { ProductCatalogDomainDTOsCategoryDto } from "./product_catalog_domain_dt_os_category_dto";
import type { ClicksiDataContractsCommonEnumsCategoryType } from "./clicksi_data_contracts_common_enums_category_type";

export interface ProductCatalogDomainDTOsCategoryTreeHistoryDetailDto {
  categories_before_changes: ProductCatalogDomainDTOsCategoryDto[];
  categories_after_changes: ProductCatalogDomainDTOsCategoryDto[];
  id: string;
  timestamp: string;
  /** @nullable */
  change_description?: string | null;
  /** @nullable */
  updated_by_user_id?: string | null;
  category_type: ClicksiDataContractsCommonEnumsCategoryType;
}